converted linkSrv.js to linkSrv.ts
